Emacs decided on hs-minor-mode
I kinda miss how vim has sane defaults, and that
you don't have to configure up every single thing.

Hideshow mode seems like a fine enough fold mode for
emacs, and it's context dependent, depending on the
major mode, which is nice. It's also available without
download.

Origami mode looks like a stroger version of this, and
might be something to consider for the future.