Readded nerdtree to git
Fzf is better in most cases, but sometimes I want an overview of files
in the current project. Also, it's easier to get to work on any system.

This commit also contains a bunch of whitespace fixes, but I couldn't be
bothered to put those into another commit, since they where so close to
the actuall data to commit.